Remove label tests.
