--- internal/engine/label/testdata/local.added.diff
@@ -1,5 +1,0 @@
---- Current
-+++ TO BE APPLIED
-@@ -0,0 +1,2 @@
-+label4; color: color6, color7
-+label5
