--- internal/engine/label/testdata/local.added.json
@@ -1,29 +1,0 @@
-[
-  {
-    "Name": "label3",
-    "Color": {
-      "Background": "color3",
-      "Text": "color4"
-    }
-  },
-  {
-    "Name": "label1"
-  },
-  {
-    "Name": "label2",
-    "Color": {
-      "Background": "color1",
-      "Text": "color2"
-    }
-  },
-  {
-    "Name": "label4",
-    "Color": {
-      "Background": "color6",
-      "Text": "color7"
-    }
-  },
-  {
-    "Name": "label5"
-  }
-]
