--- internal/engine/label/testdata/local.deleted.diff
@@ -1,4 +1,0 @@
---- Current
-+++ TO BE APPLIED
-@@ -1 +0,0 @@
--label1
