--- internal/engine/label/testdata/local.deleted.json
@@ -1,16 +1,0 @@
-[
-  {
-    "Name": "label3",
-    "Color": {
-      "Background": "color3",
-      "Text": "color4"
-    }
-  },
-  {
-    "Name": "label2",
-    "Color": {
-      "Background": "color1",
-      "Text": "color2"
-    }
-  }
-]
