--- internal/engine/label/testdata/local.modified.diff
@@ -1,5 +1,0 @@
---- Current
-+++ TO BE APPLIED
-@@ -1 +1 @@
--label1
-+label1; color: color5, color6
