--- internal/engine/label/testdata/local.modified.json
@@ -1,19 +1,0 @@
-[
-  {
-    "Name": "label3"
-  },
-  {
-    "Name": "label1",
-    "Color": {
-      "Background": "color5",
-      "Text": "color6"
-    }
-  },
-  {
-    "Name": "label2",
-    "Color": {
-      "Background": "color1",
-      "Text": "color2"
-    }
-  }
-]
