--- internal/engine/label/testdata/local.nodiff.json
@@ -1,19 +1,0 @@
-[
-  {
-    "Name": "label1"
-  },
-  {
-    "Name": "label3",
-    "Color": {
-      "Background": "color3",
-      "Text": "color4"
-    }
-  },
-  {
-    "Name": "label2",
-    "Color": {
-      "Background": "color1",
-      "Text": "color2"
-    }
-  }
-]
