--- internal/engine/label/testdata/upstream.json
@@ -1,24 +1,0 @@
-[
-  {
-    "ID": "3",
-    "Name": "label3",
-    "Color": {
-      "Background": "color3",
-      "Text": "color4"
-    }
-  },
-  {
-    "ID": "1",
-    "Name": "label1",
-    "NumMessages": 10
-  },
-  {
-    "ID": "2",
-    "Name": "label2",
-    "Color": {
-      "Background": "color1",
-      "Text": "color2"
-    },
-    "NumMessages": 3
-  }
-]
