+++ testdata/valid/03-labels.json
@@ -14,6 +14,13 @@
     },
     {
       "name": "label3"
+    },
+    {
+      "name": "label4",
+      "color": {
+        "background": "white",
+        "text": "black"
+      }
     }
   ],
   "rules": [
