+++ testdata/valid/05-bigdiff.diff
@@ -131,9 +131,11 @@ Filters:
 Labels:
 --- Current
 +++ TO BE APPLIED
-@@ -1,2 +1,3 @@
+@@ -1,3 +1,4 @@
+-label4; color: white, black
 -label2; color: red, blue
 -label3
++label4; color: white, gray
 +maillist
 +thirdlabel
 +differentlabel
